requirements.txt
# F1 Telegram Bot Requirements for Leapcell Deployment
# Core Bot Framework
python-telegram-bot==20.7

# Web Framework
flask==2.3.3
gunicorn==21.2.0
gevent==23.9.1

# HTTP Requests
requests==2.31.0
urllib3==2.0.7
httpx==0.25.2

# Environment Management
python-dotenv==1.0.0

# Web Scraping - Optimized for Leapcell
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Date/Time Handling
python-dateutil==2.8.2

# JSON Processing
orjson==3.9.10

# Async Support
asyncio-mqtt==0.16.1

# Logging
structlog==23.2.0

# System Dependencies (for playwright) - container optimized
psutil==5.9.6

# Leapcell-specific optimizations
uvloop==0.18.0; sys_platform != 'win32'
watchdog==3.0.0

# Health check and monitoring
healthcheck==1.3.3

# Additional system packages for headless browsers
Pillow==10.1.0

# Webhook handling
flask-cors==4.0.0